requirements.txt
firebase-admin>=6.5.0
requests>=2.31.0
google-generativeai>=0.8.0
playwright>=1.40.0
